Move encoders to separate files
