+++ Sources/Core/Encoding/ImageEncoders+ImageIO.swift
@@ -2,86 +2,7 @@
 //
 // Copyright (c) 2015-2021 Alexander Grebenyuk (github.com/kean).
 
-#if !os(macOS)
-import UIKit
-#else
-import Cocoa
-#endif
-
-#if os(watchOS)
-import WatchKit
-#endif
-
-import ImageIO
-
-// MARK: - ImageEncoding
-
-/// An image encoder.
-public protocol ImageEncoding {
-    /// Encodes the given image.
-    func encode(_ image: PlatformImage) -> Data?
-
-    /// An optional method which encodes the given image container.
-    func encode(_ container: ImageContainer, context: ImageEncodingContext) -> Data?
-}
-
-public extension ImageEncoding {
-    func encode(_ container: ImageContainer, context: ImageEncodingContext) -> Data? {
-        self.encode(container.image)
-    }
-}
-
-// MARK: - ImageEncoder
-
-/// Image encoding context used when selecting which encoder to use.
-public struct ImageEncodingContext {
-    public let request: ImageRequest
-    public let image: PlatformImage
-    public let urlResponse: URLResponse?
-}
-
-// MARK: - ImageEncoders
-
-/// A namespace with all available encoders.
-public enum ImageEncoders {}
-
-// MARK: - ImageEncoders.Default
-
-extension ImageEncoders {
-    /// A default adaptive encoder which uses best encoder available depending
-    /// on the input image and its configuration.
-    public struct Default: ImageEncoding {
-        public var compressionQuality: Float
-
-        /// Set to `true` to switch to HEIF when it is available on the current hardware.
-        /// `false` by default.
-        public var isHEIFPreferred = false
-
-        public init(compressionQuality: Float = 0.8) {
-            self.compressionQuality = compressionQuality
-        }
-
-        public func encode(_ image: PlatformImage) -> Data? {
-            guard let cgImage = image.cgImage else {
-                return nil
-            }
-            let type: ImageType
-            if cgImage.isOpaque {
-                if isHEIFPreferred && ImageEncoders.ImageIO.isSupported(type: .heic) {
-                    type = .heic
-                } else {
-                    type = .jpeg
-                }
-            } else {
-                type = .png
-            }
-            let encoder = ImageEncoders.ImageIO(type: type, compressionRatio: compressionQuality)
-            return encoder.encode(image)
-        }
-    }
-}
-
-// MARK: - ImageEncoders.ImageIO
+import Foundation
 
 extension ImageEncoders {
     /// An Image I/O based encoder.
