+++ Sources/Core/Encoding/ImageEncoders.swift
@@ -1,0 +1,8 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2015-2021 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+/// A namespace with all available encoders.
+public enum ImageEncoders {}
